requirements.txt
polars==0.20.25
fastexcel==0.10.4
python-dotenv==1.0.1
dash==2.17.1
plotly==5.22.0
pandas==2.2.2
websocket==0.2.1
psycopg2==2.9.9
sqlalchemy==2.0.31
connectorx==0.3.3
schedule==1.2.1
dash_bootstrap_templates==1.2.0
